requirements.txt
webdriver-manager==4.0.0
selenium==4.11.2
pytest==7.4.0
pytest-xdist==3.3.1
pytest-metadata==3.0.0
allure-pytest==2.13.2